Code Optimizations
WHAT: Make a few optimizations to the CSS code on the Clock page, add a comma to the Date string presentation.
WHY: Improve the look of the page a bit.
HOW:  made some changes to the CSS code to improve alignment of the date info, added a comma after the day of month.