Update Connection String
WHAT: The connection string had to be altered slightly.
WHY: It appears that a recent container image update for the SQL Server 2022 instance now necessitates that the port be included in the 'server' declaration .
HOW: Updated the connection string in appsettings.json so that it would make a proper connection to the SQL server.